refactor: Simplify infection process parameters by removing distance exponent and mixing scale
- Removed distance_exponent and mixing_scale from InfectionParams and TransmissionParams.
- Introduced a mixer attribute for handling mixing behavior in infection processes.
- Updated references in InfectionProcess to utilize the new mixer structure for mixing matrix calculations.
- Adjusted unit tests to align with the new parameter structure.
